layers: Create seperate SPIR-V library module
