+++ BUILD.gn
@@ -176,6 +176,8 @@ vvl_sources = [
   "layers/layer_options.h",
   "layers/object_tracker/object_lifetime_validation.h",
   "layers/object_tracker/object_tracker_utils.cpp",
+  "layers/spirv/spv_instruction.cpp",
+  "layers/spirv/spv_instruction.h",
   "layers/state_tracker/base_node.cpp",
   "layers/state_tracker/base_node.h",
   "layers/state_tracker/buffer_state.cpp",
@@ -208,8 +210,6 @@ vvl_sources = [
   "layers/state_tracker/sampler_state.h",
   "layers/state_tracker/semaphore_state.cpp",
   "layers/state_tracker/semaphore_state.h",
-  "layers/state_tracker/shader_instruction.cpp",
-  "layers/state_tracker/shader_instruction.h",
   "layers/state_tracker/shader_module.cpp",
   "layers/state_tracker/shader_module.h",
   "layers/state_tracker/shader_object_state.cpp",
