+++ layers/spirv/spv_instruction.h
@@ -24,14 +24,6 @@
 #include "containers/custom_containers.h"
 #include <spirv/unified1/spirv.hpp>
 
-struct SPIRV_MODULE_STATE;
-
-struct AtomicInstructionInfo {
-    uint32_t storage_class;
-    uint32_t bit_width;
-    uint32_t type;  // ex. OpTypeInt
-};
-
 // Holds information about a single SPIR-V instruction
 // Provides easy access to len, opcode, and content words without the caller needing to care too much about the physical SPIRV
 // module layout.
@@ -69,7 +61,6 @@ class Instruction {
     uint32_t GetConstantValue() const;
     uint32_t GetBitWidth() const;
     uint32_t GetByteWidth() const { return (GetBitWidth() + 31) / 32; }
-    AtomicInstructionInfo GetAtomicInfo(const SPIRV_MODULE_STATE& module_state) const;
     spv::BuiltIn GetBuiltIn() const;
 
     bool IsArray() const;
