+++ scripts/generators/spirv_grammar_generator.py
@@ -257,7 +257,7 @@ class SpirvGrammarHelperOutputGenerator(BaseGenerator):
         out.append('''
             #include "containers/custom_containers.h"
             #include "spirv_grammar_helper.h"
-            #include "state_tracker/shader_instruction.h"
+            #include "spirv/spv_instruction.h"
 
             // All information related to each SPIR-V opcode instruction
             struct InstructionInfo {
